Changed demand to double and introduced demand scaling

+++ Applications/CNRP/include/cnrp_cg.h
@@ -29,21 +29,21 @@
 
 typedef struct CG_CNRP_SPEC{
    cnrp_cg_params par;
-   int           dg_id;   /*contains the tid of the graphics window*/
-   int           vertnum;  /*the number of nodes in the problem,
+   int           dg_id;    /* contains the tid of the graphics window*/
+   int           vertnum;  /* the number of nodes in the problem,
 			      including the depot                */
-   int          *demand;   /* alist of the customer demands*/
-   int           capacity; /*the capacity of the trucks*/
-   int           numroutes;/*contains the number of routes that the problem
+   double       *demand;   /* a list of the customer demands*/
+   double        capacity; /* the capacity of the trucks*/
+   int           numroutes;/* contains the number of routes that the problem
 			      is to be solved with. can be prespecified.  */
-   int          *edges;    /*contains a list of the edges in the current
+   int          *edges;    /* contains a list of the edges in the current
 			      subproblem*/
    network      *n;
    int           orig_edgenum;
    int          *cost;
    int          *ref;      /* the last five  are for the shrinking routines; */
-   char         *in_set;   /* They are here to optimize/speed up things */
-   int          *new_demand;
+   char         *in_set;   /* they are here to optimize/speed up things */
+   double       *new_demand;
    double       *cut_val;
    char         *cut_list;
 
@@ -71,53 +71,53 @@ typedef struct CG_CNRP_SPEC{
 /*========================= Other user subroutines =========================*/
 /*===========================================================================*/
 
-int check_connectivity PROTO((network *n, double etol, int capacity,
+int check_connectivity PROTO((network *n, double etol, double capacity,
 			      int numroutes, char mult));
 
-int check_flow_connectivity PROTO((network *n, double etol, int capacity,
+int check_flow_connectivity PROTO((network *n, double etol, double capacity,
 				   int numroutes, char mult));
 
 /*===========================================================================*/
 /*=============================== shrink.c ==================================*/
 /*===========================================================================*/
 
-int reduce_graph PROTO((network *n, double etol, int *demand, int capacity,
-			 int mult, cut_data *new_cut));
+int reduce_graph PROTO((network *n, double etol, double *demand,
+			double capacity, int mult, cut_data *new_cut));
 int greedy_shrinking1 PROTO((network *n, double truck_cap, double etol,
 			     int max_num_cuts, cut_data *new_cut,
 			     int *compnodes, int *compmembers, int compnum,
 			     char *in_set, double *cut_val,int *ref,
-			     char *cut_list, int *demand, int mult));
+			     char *cut_list, double *demand, int mult));
 int greedy_shrinking6 PROTO((network *n, double truck_cap,
 			     double etol, cut_data *new_cut,
 			     int *compnodes,
 			     int *compmembers, int compnum, char *in_set,
 			     double *cut_val,int *ref, char *cut_list,
-			     int max_num_cuts, int *demand, int trial_num,
+			     int max_num_cuts, double *demand, int trial_num,
 			     double prob, int mult));
 int greedy_shrinking1_one PROTO((network *n, double truck_cap,
 				 double etol, int max_num_cuts,
 				 cut_data *new_cut, char *in_set,
 				 double *cut_val, char *cut_list,
-				 int num_routes, int *demand, int mult));
+				 int num_routes, double *demand, int mult));
 int greedy_shrinking6_one PROTO((network *n, double truck_cap,
 				 double etol, cut_data *new_cut,
 				 char *in_set, double *cut_val, int num_routes,
 				 char *cut_list, int max_num_cuts,
-				 int *demand,int trial_num, double prob,
+				 double *demand,int trial_num, double prob,
 				 int mult));
 int greedy_shrinking2_one PROTO((network *n, double truck_cap,
 				 double etol, cut_data *new_cut,
 				 char *in_set, double *cut_val, int num_routes,
-				 int *demand, int mult));
+				 double *demand, int mult));
 
 /*===========================================================================*/
 /*============================ biconnected.c ================================*/
 /*===========================================================================*/
 
 void depth_first_search PROTO((vertex *v, int *count1, int *count2));
-int biconnected PROTO((network *n, int *compnodes, int
-			   *compdemands, double *compcuts));
+int biconnected PROTO((network *n, int *compnodes, double *compdemands,
+		       double *compcuts));
 void compute_comp_nums PROTO((vertex *v, int parent_comp, int *num_comps,
 		       char parent_is_art_point));
 int tsp_cuts PROTO((network *n, int verbosity, char tsp_prob, int which_cuts));
